feat: 스켈레톤 UI 로딩 수정, 장소검색 컴포넌트 분리
